add themes.js module
